Remove the reference for binaryen.Module for DataSegmentContext
BUG=N/A
